Stabilize instrumentation-api and introduce jApiCmp checks (#6566)
* Stabilize instrumentation-api and introduce jApiCmp checks

* Update
