+++ instrumentation-api/gradle.properties
@@ -1,0 +1,1 @@
+otel.stable=true
